/advancedEnemies:
Infinite FPS bug fixed
